+++ flaredb/src/engine/mod.rs
@@ -1,4 +1,3 @@
-pub mod coders;
 pub mod dispatcher;
 pub mod executor;
 pub mod harness;
