+++ flaredb/src/lib.rs
@@ -58,6 +58,7 @@
 //! Project repository:
 //!
 //! https://github.com/flare-db/flare-db
+pub mod coders;
 pub mod engine;
 pub mod fusion;
 pub mod io;
